Improved the syntax of my fake pattern language, in addition to adapting the semantics of @sts.
Variables must now start with '@' and variable names should be in lowercase letters e.g. '@General' has become '@general'.

For @sts --development status, the 'finished' status is renamed to 'closed'.

The semantics of elements are expressed in gentext.xsl --babelfish.

Index-wise, such indices are now generated:

Resolved Requirements
  Implemented
    Foo
Open Requirements
  Raised
     Bar

In general, the @sts represents the state of a requirement that is NOT resolved yet.

That's why @general evaluates to 'Open' for @sts values.

There's an exception though. If @sts='finished', the requirement is not open anymore.

Improved the syntax of my fake pattern language, in addition to adapting the semantics of @sts.

Variables must now start with '@' and variable names should be in lowercase letters e.g. '@General' has become '@general'.

For @sts --development status, the 'finished' status is renamed to 'closed'.

The semantics of elements are expressed in gentext.xsl --babelfish.

Index-wise, such indices are now generated:

Resolved Requirements
  Implemented
    Foo
Open Requirements
  Raised
     Bar

In general, the @sts represents the state of a requirement that is NOT resolved yet.

That's why @general evaluates to 'Open' for @sts values.

There's an exception though. If @sts='finished', the requirement is not open anymore.

This explains why the general term (only) for 'closed' --formerly known as 'finished'-- evaluates to 'Closed'.

Bind @sts, @res, and @pty in TeX after the resolution of their values usinImproved the syntax of my fake pattern language, in addition to adapting the semantics of @sts.

Variables must now start with '@' and variable names should be in lowercase letters e.g. '@General' has become '@general'.

For @sts --development status, the 'finished' status is renamed to 'closed'.

The semantics of elements are expressed in gentext.xsl --babelfish.

Index-wise, such indices are now generated:

Resolved Requirements
  Implemented
    Foo
Open Requirements
  Raised
     Bar

In general, the @sts represents the state of a requirement that is NOT resolved yet.

That's why @general evaluates to 'Open' for @sts values.

There's an exception though. If @sts='finished', the requirement is not open anymore.

This explains why the general term (only) for 'closed' --formerly known as 'finished'-- evaluates to 'Closed'.

Bind @sts, @res, and @pty in TeX after the resolution of their values using gentext.
g gentext.
This explains why the general term (only) for 'closed' --formerly known as 'finished'-- evaluates to 'Closed'.

Bind @sts, @res, and @pty in TeX after the resolution of their values usinImproved the syntax of my fake pattern language, in addition to adapting the semantics of @sts.

Variables must now start with '@' and variable names should be in lowercase letters e.g. '@General' has become '@general'.

For @sts --development status, the 'finished' status is renamed to 'closed'.

The semantics of elements are expressed in gentext.xsl --babelfish.

Index-wise, such indices are now generated:

Resolved Requirements
  Implemented
    Foo
Open Requirements
  Raised
     Bar

In general, the @sts represents the state of a requirement that is NOT resolved yet.

That's why @general evaluates to 'Open' for @sts values.

There's an exception though. If @sts='finished', the requirement is not open anymore.

This explains why the general term (only) for 'closed' --formerly known as 'finished'-- evaluates to 'Closed'.

Bind @sts, @res, and @pty in TeX after the resolution of their values using gentext.
g gentext.