Improved the syntax of my fake pattern language, in addition to adapting the semantics of @sts.
Index-wise, such indices are now generated:

Resolved Requirements
  Implemented
    Foo
Open Requirements
  Raised
     Bar

In the pattern language, variables must now start with '@' and variable names should be in lowercase letters e.g. '@General' has become '@general'.

For @sts --development status, the 'finished' status is renamed to 'closed'.

NOTE: The semantics of elements are expressed in gentext.xsl --babelfish.

In general, the @sts represents the state of a requirement that is NOT resolved yet.

That's why @general evaluates to 'Open' for @sts states.

There's an exception though. If @sts='closed', then the requirement is not open anymore.

This explains why the general term (only) for 'closed' --formerly known as 'finished'-- evaluates to 'Resolved'.

Changes in bindings.xsl:
Bind @sts, @res, and @pty in TeX after the resolution of their values using gentext.