Fixed patterns that were affected after adapting the semantics of srs:in in order to provide a syntactic sugar --see below-- for IO documentation abstractions i.e. srs:in[@name] should be srs:in[not(@ref)].
This because we should count <in/> and <in name="..."/>s in order to match all the previously defined input operations. Clearly, srs:in[@name] doesn't match srs:in. That's why we use srs:in[not(@ref)] in order to find all input definitions.

The same is true for output operations.

The syntactic sugar allows one to not have to define a name --local identifier-- for an IO operation. An example:

<ols>
<itm>User enters their <in>username</in></itm>
<itm>My program checks whether <lin/> already exists</itm>
</ols>

is equivalent to:

<ols>
<itm>User enters their <in name="uname">username</in></itm>
<itm>My program checks whether <in ref="uname"/> already exists</itm>
</ols>